Ver01/requirements.txt
streamlit==1.28.1
psycopg2-binary==2.9.7
supabase==2.19.0
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0
pandas==2.1.1
openpyxl==3.1.2
reportlab==4.0.4
requests==2.31.0
plotly==5.16.1
python-dateutil==2.8.2